requirements.txt
# Speculator Bot - AI-Assisted Predictive Analysis Tool
# Core dependencies

# ML and Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Database support
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.5.0

# Git operations
gitpython>=3.1.40

# Code analysis
ast-comments>=1.1.0
radon>=6.0.1

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI
click>=8.1.0
rich>=13.0.0
tabulate>=0.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
tqdm>=4.66.0

# Model Context Protocol (MCP)
mcp>=0.1.0

# SAP AI Core Integration
ai-core-sdk>=2.6.0

